+++ vkconfig_core/configuration.h
@@ -43,8 +43,7 @@ class Configuration {
 
     std::string key;  // User readable display of the configuration name (may contain spaces)
     int platform_flags;
-    std::string description;        // A friendly description of what this profile does
-    QByteArray setting_tree_state;  // Recall editor tree state
+    std::string description;  // A friendly description of what this profile does
 
     std::vector<Parameter> parameters;