+++ vkconfig_core/parameter.h
@@ -40,7 +40,7 @@ struct Parameter {
 
     Parameter() : state(LAYER_STATE_APPLICATION_CONTROLLED), platform_flags(PLATFORM_ALL_BIT), overridden_rank(NO_RANK) {}
     Parameter(const std::string& key, const LayerState state)
-        : key(key), state(state), platform_flags(PLATFORM_ALL_BIT), overridden_rank(NO_RANK) {}
+        : key(key), state(state), platform_flags(PLATFORM_ALL_BIT), overridden_rank(NO_RANK), collapsed(false) {}
 
     bool ApplyPresetSettings(const LayerPreset& preset);
 
@@ -49,6 +49,7 @@ struct Parameter {
     int platform_flags;
     SettingDataSet settings;
     int overridden_rank;
+    bool collapsed;
 };
 
 ParameterRank GetParameterOrdering(const std::vector<Layer>& available_layers, const Parameter& parameter);